Created log sink for E2E tests using serilog

+++ test/Dapr.E2E.Test/Workflows/WorkflowTest.cs
@@ -43,12 +43,13 @@ namespace Dapr.E2E.Test
             var health = await daprClient.CheckHealthAsync();
             health.Should().Be(true, "DaprClient is not healthy");
 
-            var searchTask = Task.Run(async() =>
+            var searchTask = Task.Run(() =>
             {
                 using (StreamReader reader = new StreamReader(logFilePath))
                 {
                     string line;
-                    while ((line = await reader.ReadLineAsync().WaitAsync(cts.Token)) != null)
+
+                    while ((line = reader.ReadLine()) != null)
                     {
                         foreach (var entry in logStrings)
                         {
@@ -66,18 +67,12 @@ namespace Dapr.E2E.Test
                 }
             }, cts.Token);
 
-            try
-            {
-                await searchTask;
-            }
-            finally
+            if (!Task.WaitAll(new Task[] {searchTask}, timeout))
             {
                 File.Delete(logFilePath);
+                Assert.True(false, "Expected logs weren't found within timeout period");
             }
-            if (!allLogsFound)
-            {
-                Assert.True(false, "The logs were not able to found within the timeout");
-            }
+            File.Delete(logFilePath);
         }
         [Fact]
         public async Task TestWorkflows()
